+++ src/lib/api_elements/auth.js
@@ -1,18 +1,18 @@
 var oauthRegex = /access_token=([^&]+)/;
 var oauthDeniedRegex = /error=access_denied/;
 
-
 var promises = require("q");
 var helpers = require('../reusables/helpers');
 var dom = require('../reusables/dom');
 var messages = require('../reusables/messages');
 var errorify = require("./errorify");
 
-
 var ENDPOINTS_userinfo = require("../enum/endpoints").userinfo;
+var ENDPOINTS_tokenauth = require("../enum/endpoints").tokenauth;
 
 
 function Auth(options) {
+    this.type = "Bearer";
     this.options = options;
     if (this.options.token) {
         this.token = this.options.token;
@@ -25,7 +25,7 @@ function Auth(options) {
 var authPrototypeMethods = {};
 
 authPrototypeMethods._reloadForToken = function () {
-    window.location.href = this.options.egnyteDomainURL + "/puboauth/token?client_id=" + this.options.key + "&mobile=" + ~~(this.options.mobile) + "&redirect_uri=" + window.location.href;
+    window.location.href = this.options.egnyteDomainURL + ENDPOINTS_tokenauth + "?client_id=" + this.options.key + "&mobile=" + ~~(this.options.mobile) + "&redirect_uri=" + window.location.href;
 }
 
 authPrototypeMethods._checkTokenResponse = function (success, denied, notoken, overrideWindow) {
@@ -61,8 +61,9 @@ authPrototypeMethods.requestTokenIframe = function (targetNode, callback, denied
     if (!this.token) {
         var self = this;
         var locationObject = window.location;
+
         emptyPageURL = (emptyPageURL) ? locationObject.protocol + "//" + locationObject.host + emptyPageURL : locationObject.href;
-        var url = this.options.egnyteDomainURL + "/puboauth/token?client_id=" + this.options.key + "&mobile=" + ~~(this.options.mobile) + "&redirect_uri=" + emptyPageURL;
+        var url = self.options.egnyteDomainURL + ENDPOINTS_tokenauth + "?client_id=" + self.options.key + "&mobile=" + ~~(self.options.mobile) + "&redirect_uri=" + emptyPageURL;
         var iframe = dom.createFrame(url, !!"scrollbars please");
         iframe.onload = function () {
             try {
@@ -89,7 +90,6 @@ authPrototypeMethods.requestTokenIframe = function (targetNode, callback, denied
     } else {
         callback();
     }
-
 }
 
 
@@ -109,10 +109,11 @@ authPrototypeMethods._postTokenUp = function () {
 
     }
 }
+
 authPrototypeMethods.requestTokenPopup = function (callback, denied, recvrURL) {
     var self = this;
     if (!this.token) {
-        var url = this.options.egnyteDomainURL + "/puboauth/token?client_id=" + this.options.key + "&mobile=" + ~~(this.options.mobile) + "&redirect_uri=" + recvrURL;
+        var url = this.options.egnyteDomainURL + ENDPOINTS_tokenauth + "?client_id=" + this.options.key + "&mobile=" + ~~(this.options.mobile) + "&redirect_uri=" + recvrURL;
         var win = window.open(url);
         win.name = this.options.channelMarker;
         var handler = messages.createMessageHandler(null, this.options.channelMarker, function (message) {
@@ -138,7 +139,7 @@ authPrototypeMethods.requestTokenByPassword = function (username, password) {
 
     return this.requestEngine.promiseRequest({
         method: "POST",
-        url: this.options.egnyteDomainURL + "/puboauth/token",
+        url: this.options.egnyteDomainURL + ENDPOINTS_tokenauth + "",
         headers: {
             "content-type": "application/x-www-form-urlencoded"
         },
@@ -148,7 +149,7 @@ authPrototypeMethods.requestTokenByPassword = function (username, password) {
             "username=" + username,
             "password=" + password
         ].join("&")
-    },null,!!"forceNoAuth").then(function (result) { //result.response result.body
+    }, null, !!"forceNoAuth").then(function (result) { //result.response result.body
         self.token = result.body.access_token
         return self.token;
     });