+++ src/lib/api_elements/notes.js
@@ -10,14 +10,11 @@ exports.addNote = function (pathFromRoot, body) {
         pathFromRoot = helpers.encodeNameSafe(pathFromRoot);
         var opts = {
             method: "POST",
-            headers: {
-                "content-type": "application/vnd.egnyte.annotations.request+json;v=1"
-            },
             url: requestEngine.getEndpoint() + ENDPOINTS_notes,
-            body: JSON.stringify({
+            json:{
                 "path": pathFromRoot,
                 "body": body,
-            })
+            }
         };
         return requestEngine.promiseRequest(decorate(opts));
     }).then(function (result) { //result.response result.body