+++ src/lib/api_elements/storage.js
@@ -13,7 +13,7 @@ function Storage(requestEngine) {
 }
 
 var storageProto = {};
-storageProto.exists = function (pathFromRoot, versionEntryId) {
+storageProto.exists = function (pathFromRoot) {
     var requestEngine = this.requestEngine;
     var decorate = this.getDecorator();
     return promises(true).then(function () {
@@ -22,12 +22,6 @@ storageProto.exists = function (pathFromRoot, versionEntryId) {
             method: "GET",
             url: requestEngine.getEndpoint() + ENDPOINTS.fsmeta + encodeURI(pathFromRoot),
         };
-
-        if (versionEntryId) {
-            opts.params = {
-                "entry_id": versionEntryId
-            };
-        }
 
         return requestEngine.promiseRequest(decorate(opts));
     }).then(function (result) { //result.response result.body
@@ -109,7 +103,6 @@ storageProto.createFolder = function (pathFromRoot) {
     }).then(function (result) { //result.response result.body
         if (result.response.statusCode == 201) {
             return {
-                id: result.response.headers["etag"],
                 path: pathFromRoot
             };
         }
@@ -136,7 +129,7 @@ function transfer(requestEngine, decorate, pathFromRoot, newPath, action) {
             url: requestEngine.getEndpoint() + ENDPOINTS.fsmeta + encodeURI(pathFromRoot),
             json: {
                 "action": action,
-                "destination": "/" + newPath,
+                "destination": newPath,
             }
         };
         return requestEngine.promiseRequest(decorate(opts));
@@ -265,9 +258,9 @@ storageProto.removeFileVersion = function (pathFromRoot, versionEntryId) {
 }
 
 
-storageProto.remove = function (pathFromRoot) {
+storageProto.remove = function (pathFromRoot, versionEntryId) {
     var decorate = this.getDecorator();
-    return remove(this.requestEngine, decorate, pathFromRoot);
+    return remove(this.requestEngine, decorate, pathFromRoot, versionEntryId);
 }
 
 storageProto = helpers.extend(storageProto,notes);